Scott Oonk's for for x86_64 compile warnings.
git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/trunk@11941 57a11ea4-9604-0410-9ed3-97b8803252fd

+++ src/gnome-utils/gnc-main-window.c
@@ -620,7 +620,7 @@ gnc_main_window_restore_all_state (gpointer session, gpointer unused)
   /*  Debugging: dump a copy to stdout and the trace log */
   {
     gchar *file_data;
-    gint file_length;
+    gsize file_length;
     file_data = g_key_file_to_data(data.key_file, &file_length, NULL);
     DEBUG("=== File Data Read===\n%s\n=== File End ===\n", file_data);
     g_free(file_data);
@@ -820,7 +820,7 @@ gnc_main_window_save_all_state (gpointer session, gpointer user_data)
 	/*  Debugging: dump a copy to the trace log */
 	{
 	  gchar *file_data;
-	  gint file_length;
+	  gsize file_length;
 	  file_data = g_key_file_to_data(data.key_file, &file_length, NULL);
 	  DEBUG("=== File Data Written===\n%s\n=== File End ===\n", file_data);
 	  g_free(file_data);
